Move signal handling code to pkg/signal.Trap
